! relevance param is a float not an int. (Seatch.php)
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@6734 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ Sources/Search.php
@@ -1603,7 +1603,7 @@ function PlushSearch2()
 						{
 							$smcFunc['db_insert']('',
 								'{db_prefix}log_search_results',
-								array('id_search' => 'int', 'id_topic' => 'int', 'relevance' => 'int', 'id_msg' => 'int', 'num_matches' => 'int'),
+								array('id_search' => 'int', 'id_topic' => 'int', 'relevance' => 'float', 'id_msg' => 'int', 'num_matches' => 'int'),
 								$inserts,
 								array('id_search', 'id_topic')
 							);
